requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async==0.6.2
async-timeout==4.0.2
attrs==22.1.0
binance==0.3
certifi==2022.6.15.1
charset-normalizer==2.1.1
dateparser==1.1.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
python-binance==1.0.16
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
